LayoutTests/editing/pasteboard/paste-table-001.html

<html> 
<head>
<style>
.editing { 
    border: 2px solid red; 
    padding: 12px; 
    font-size: 24px; 
}
</style>
<script src=../editing.js language="JavaScript" type="text/JavaScript" ></script>
<script src="../../resources/dump-as-markup.js"></script>
<title>Editing Test</title> 
</head> 
<body>
<div contenteditable id="root" class="editing">
<span id="test">Empty table cell</span>
<p>
<table border="2">
<tr>
    <td>Left column</td>
    <td>Right column</td>
</tr>
<tr>
    <td></td>
    <td>I should be in the right column.</td>
</tr>
</table>
</div>
<script>

function editingTest() {
    moveSelectionForwardByLineCommand();
    for (i = 0; i < 3; i++)
        extendSelectionForwardByLineCommand();
    for (i = 0; i < 32; i++)
        extendSelectionForwardByCharacterCommand();
    copyCommand();
    moveSelectionForwardByLineCommand();
    pasteCommand();
    Markup.dump('root');
}

runDumpAsTextEditingTest(true);

</script>
</body>
</html>
